Procarea: gear premio, scudi per ruolo, revisione IT e fix spawn/dispel
Estende le istanze procarea con equip premio generato (boot + roll resistenze),
scudi con modello tank/caster/hybrid in base al party, e AFF_FLYING sui mob in
stanze SECT_AIR per evitare cadute spurie.
Revisione italiana sistematica di mob, temi e descrizioni stanze (minimo 4 righe)
via procarea_room_desc.py e rigenerazione degli .inc.
Corregge il log bash/set skill (ubyte → int) e allinea il bonus dispel spellpower
ai caster puri (IS_CASTER_N), inclusi multi cleric/mage.
Aggiunge cooldown uscita Darkstar da istanza attiva e ritocchi testi vortice solo.

+++ src/act.off.cpp
@@ -1588,12 +1588,12 @@ ACTION_FUNC(do_bash) {
 		percent = number(10,60);
 		/*
 		 mudlog( LOG_CHECK,   "percent for 1st bash %d vs %d", percent,
-		       ch->skills[ SKILL_BASH ].learned );
+		       static_cast<int>(ch->skills[ SKILL_BASH ].learned));
 		*/
 	}
 	else {
 		mudlog(LOG_CHECK, "percent for    bash %d vs %d", percent,
-			   ch->skills[ SKILL_BASH ].learned);
+			   static_cast<int>(ch->skills[ SKILL_BASH ].learned));
 	}
 
 
@@ -1611,7 +1611,7 @@ ACTION_FUNC(do_bash) {
 	percent+=MAX(30,MAX(-5,pesi));
 #endif
 	/*  mudlog( LOG_CHECK, "percent for bash adj %d vs %d", percent,
-	          ch->skills[ SKILL_BASH ].learned );
+	          static_cast<int>(ch->skills[ SKILL_BASH ].learned));
 	 */
 	if(percent > MIN(100, ch->skills[ SKILL_BASH ].learned)) {
 		if(GET_POS(victim) > POSITION_DEAD) {
